backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.2
sqlalchemy==2.0.23
PyPDF2==3.0.1
pdfplumber==0.10.3
numpy
scikit-learn
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1
python-dotenv==1.0.0
cryptography==41.0.7

